Fundamental Kotlin - Small cleanup.
